feat: add combinatorial encoding and fix 32-bit overflow
Z1 Milestone: Combinatorial Number System for RRR compression
- Add internal/combinatorial.go with binomial table, CombEncode, CombDecode, OffsetBits
- Add exhaustive tests covering all 32768 15-bit patterns
- Performance: ~10ns encode, ~12ns decode, zero allocations

Fix 32-bit platform overflow:
- Use uint64(1) << offset instead of 1 << offset in bit shift operations
- Prevents overflow when shift amount >= 32 on 32-bit systems

+++ internal/bitops.go
@@ -13,7 +13,7 @@ func Popcount(x uint64) int {
 func SelectInBlock(block uint64, rank int) int {
 	count := 0
 	for i := 0; i < 64; i++ {
-		if (block & (1 << i)) != 0 {
+		if (block & (uint64(1) << i)) != 0 {
 			count++
 			if count == rank {
 				return i
